feat: add Persistence skeleton for db writes
