+++ crates/engine/tree/src/lib.rs
@@ -18,9 +18,9 @@ pub mod chain;
 pub mod download;
 /// Engine Api chain handler support.
 pub mod engine;
+/// The background writer task for batch db writes.
+pub mod persistence;
 /// Support for managing the pipeline.
 pub mod pipeline;
 /// Support for interacting with the blockchain tree.
 pub mod tree;
-/// The background writer task for batch db writes.
-pub mod background_writer;
